refactor: split bonus apply handlers into bonus modules
