+++ src/crimson/bonuses/weapon.py
@@ -1,0 +1,19 @@
+from __future__ import annotations
+
+from ..perks import PerkId
+from .apply_context import BonusApplyCtx
+
+
+def apply_weapon(ctx: BonusApplyCtx) -> None:
+    from ..gameplay import perk_active, weapon_assign_player
+
+    weapon_id = int(ctx.amount)
+    if perk_active(ctx.player, PerkId.ALTERNATE_WEAPON) and ctx.player.alt_weapon_id is None:
+        ctx.player.alt_weapon_id = int(ctx.player.weapon_id)
+        ctx.player.alt_clip_size = int(ctx.player.clip_size)
+        ctx.player.alt_ammo = float(ctx.player.ammo)
+        ctx.player.alt_reload_active = bool(ctx.player.reload_active)
+        ctx.player.alt_reload_timer = float(ctx.player.reload_timer)
+        ctx.player.alt_shot_cooldown = float(ctx.player.shot_cooldown)
+        ctx.player.alt_reload_timer_max = float(ctx.player.reload_timer_max)
+    weapon_assign_player(ctx.player, weapon_id, state=ctx.state)
